Form_Handling/requirements.txt
click>=8.1.3
et-xmlfile==1.1.0
Jinja2>=3.1.2
MarkupSafe==2.1.1
openpyxl==3.0.9
Pillow==8.4.0
Werkzeug>=2.3.7